--- src/main/java/com/programmers/domain/Food.java
@@ -1,56 +1,0 @@
-package com.programmers.domain;
-
-import org.hibernate.annotations.DynamicUpdate;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@DynamicUpdate
-public class Food {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column(length = 30, nullable = false)
-    private String name;
-
-    @Column(length = 30, nullable = false)
-    private int price;
-
-    @Column(columnDefinition = "TEXT")
-    private String description;
-
-    private String image;
-
-
-    @Builder
-    public Food(Long id, String name, int price, String description, String image) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.description = description;
-        this.image = image;
-    }
-
-    public void changeName(String name) {
-        this.name = name;
-    }
-
-    public void changePrice(int price) {
-        this.price = price;
-    }
-
-    public void changeDescription(String description) {
-        this.description = description;
-    }
-}
