+++ src/main/java/com/programmers/domain/user/User.java
@@ -1,0 +1,100 @@
+package com.programmers.domain.user;
+
+import org.hibernate.Hibernate;
+import org.hibernate.annotations.DynamicUpdate;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+
+import java.time.LocalDateTime;
+import java.util.Objects;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.validation.constraints.NotBlank;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@DynamicUpdate
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long userId;
+
+    @NotBlank(message = "이름은 필수 입력 값입니다.")
+    @Column(nullable = false)
+    private String name;
+
+    @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
+    @Column(nullable = false)
+    private String password;
+
+    @NotBlank(message = "닉네임은 필수 입력 값입니다.")
+    @Column(nullable = false, unique = true)
+    private String nickName;
+    private String grade;
+
+    @NotBlank(message = "전화번호는 필수 입력 값입니다.")
+    @Column(nullable = false, unique = true)
+    private String phoneNumber;
+    private String role;
+
+
+    @CreatedDate
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    private LocalDateTime updatedAt;
+
+    @Builder
+    public User(Long userId, String name, String password, String nickName, String grade, String phoneNumber, String role, LocalDateTime createdAt, LocalDateTime updatedAt) {
+        this.userId = userId;
+        this.name = name;
+        this.password = password;
+        this.nickName = nickName;
+        this.grade = grade;
+        this.phoneNumber = phoneNumber;
+        this.role = role;
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+
+
+    public void changeName(String name) {
+        this.name = name;
+    }
+
+    public void changeNickName(String nickName) {
+        this.nickName = nickName;
+    }
+
+    public void changePassword(String password) {
+        this.password = password;
+    }
+
+    public void changePhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
+        User user = (User) o;
+        return userId != null && Objects.equals(userId, user.userId);
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
+}
