--- src/main/java/com/programmers/exception/DuplicatePhoneNumberException.java
@@ -1,7 +1,0 @@
-package com.programmers.exception;
-
-public class DuplicatePhoneNumberException extends RuntimeException{
-    public DuplicatePhoneNumberException() {
-        super("중복된 전화번호입니다.");
-    }
-}
