--- src/main/java/com/programmers/exception/FoodNotFoundException.java
@@ -1,7 +1,0 @@
-package com.programmers.exception;
-
-public class FoodNotFoundException extends RuntimeException{
-    public FoodNotFoundException(){
-        super("해당 id의 음식이 존재하지 않습니다.");
-    }
-}
