--- src/main/java/com/programmers/exception/InvalidTokenException.java
@@ -1,7 +1,0 @@
-package com.programmers.exception;
-
-public class InvalidTokenException extends RuntimeException {
-    public InvalidTokenException() {
-        super("인증되지 않은 토큰입니다.");
-    }
-}
