--- src/main/java/com/programmers/exception/JwtTokenProviderException.java
@@ -1,7 +1,0 @@
-package com.programmers.exception;
-
-public class JwtTokenProviderException extends RuntimeException {
-    public JwtTokenProviderException(String message) {
-        super(message);
-    }
-}
