--- src/main/java/com/programmers/exception/LoginFailedException.java
@@ -1,7 +1,0 @@
-package com.programmers.exception;
-
-public class LoginFailedException extends RuntimeException {
-    public LoginFailedException() {
-        super("로그인에 실패하였습니다.");
-    }
-}
