--- src/main/java/com/programmers/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.programmers.exception;
-
-public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException() {
-        super("해당 사용자가 없습니다.");
-    }
-}
