+++ src/main/java/com/programmers/exception/food/FoodNotFoundException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.food;
+
+public class FoodNotFoundException extends RuntimeException{
+    public FoodNotFoundException(){
+        super("해당 id의 음식이 존재하지 않습니다.");
+    }
+}
