+++ src/main/java/com/programmers/exception/menu/MenuNotFoundException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.menu;
+
+public class MenuNotFoundException extends RuntimeException{
+    public MenuNotFoundException(){
+        super("해당 id의 메뉴가 존재하지 않습니다.");
+    }
+}
