+++ src/main/java/com/programmers/exception/security/JwtTokenProviderException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.security;
+
+public class JwtTokenProviderException extends RuntimeException {
+    public JwtTokenProviderException(String message) {
+        super(message);
+    }
+}
