+++ src/main/java/com/programmers/exception/security/LoginFailedException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.security;
+
+public class LoginFailedException extends RuntimeException {
+    public LoginFailedException() {
+        super("로그인에 실패하였습니다.");
+    }
+}
