+++ src/main/java/com/programmers/exception/user/DuplicateNickNameException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.user;
+
+public class DuplicateNickNameException extends RuntimeException{
+    public DuplicateNickNameException(){
+        super("중복된 이메일입니다.");
+    }
+}
