+++ src/main/java/com/programmers/exception/user/DuplicatePhoneNumberException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.user;
+
+public class DuplicatePhoneNumberException extends RuntimeException{
+    public DuplicatePhoneNumberException() {
+        super("중복된 전화번호입니다.");
+    }
+}
