+++ src/main/java/com/programmers/exception/user/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.programmers.exception.user;
+
+public class UserNotFoundException extends RuntimeException{
+    public UserNotFoundException() {
+        super("해당 사용자가 없습니다.");
+    }
+}
